# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NutritionData {
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
  fiber?: number;
  sugar?: number;
}

export interface FoodItem {
  name: string;
  portion?: string;
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
  fiber?: number;
  sugar?: number;
}

export interface Meal {
  id: number;
  date: string;
  time: string;
  imageUrl?: string | null;
  notes?: string | null;
  foodItems: FoodItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFats: number;
  healthScore?: number | null;
}

export interface LogMealRequest {
  date: string;
  time: string;
  imageUrl?: string | null;
  notes?: string | null;
  foodItems: FoodItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFats: number;
  healthScore?: number | null;
}

export interface AnalyzeImageRequest {
  imageBase64: string;
}

export interface AnalyzeImageResponse {
  foodItems: FoodItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFats: number;
  healthScore: number;
  description: string;
}

export interface DailySummary {
  date: string;
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFats: number;
  meals: Meal[];
  calorieGoal: number;
  proteinGoal: number;
  carbGoal: number;
  fatGoal: number;
}

export interface DayData {
  date: string;
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
}

export interface WeeklySummary {
  startDate: string;
  endDate: string;
  days: DayData[];
  averageCalories: number;
  averageProtein: number;
  averageCarbs: number;
  averageFats: number;
}

export interface MonthlySummary {
  year: number;
  month: number;
  days: DayData[];
  averageCalories: number;
  averageProtein: number;
  averageCarbs: number;
  averageFats: number;
}

export interface UserSettings {
  id: number;
  dailyCalorieGoal: number;
  dailyProteinGoal: number;
  dailyCarbGoal: number;
  dailyFatGoal: number;
}

export interface UpdateSettingsRequest {
  dailyCalorieGoal: number;
  dailyProteinGoal: number;
  dailyCarbGoal: number;
  dailyFatGoal: number;
}

export type GetMealsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};

export type GetDailySummaryParams = {
  date: string;
};

export type GetWeeklySummaryParams = {
  startDate: string;
};

export type GetMonthlySummaryParams = {
  year: number;
  month: number;
};