# lib/api-zod/src/generated/types/dayData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DayData {
  date: string;
  calories: number;
  protein: number;
  carbs: number;
  fats: number;
}