# lib/api-zod/src/generated/types/logMealRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FoodItem } from "./foodItem";

export interface LogMealRequest {
  date: string;
  time: string;
  imageUrl?: string | null;
  notes?: string | null;
  foodItems: FoodItem[];
  totalCalories: number;
  totalProtein: number;
  totalCarbs: number;
  totalFats: number;
  healthScore?: number | null;
}