# lib/api-zod/src/generated/types/monthlySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DayData } from "./dayData";

export interface MonthlySummary {
  year: number;
  month: number;
  days: DayData[];
  averageCalories: number;
  averageProtein: number;
  averageCarbs: number;
  averageFats: number;
}