# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSettings {
  id: number;
  dailyCalorieGoal: number;
  dailyProteinGoal: number;
  dailyCarbGoal: number;
  dailyFatGoal: number;
}